requirements/core.txt
click
toml
pyyaml
alive-progress==3.1.4
